Sad attempt at backend for cart :(
I tried to do the backend

So basically I just added to the mainpage servelet.

Key note:
I changed the button to return the barcode rather than the ID coz i saw that there was a way to get the product with just the product but I didn't see if there was a way to search by just the ID yet. I could be wrong about this but it's just temporary and doesn't even work anyways. Just to keep in mind.