Add to cart functionality: Doesn't work
This is an attempt at making making the add to cart thing work and have the cartlist  show up on the cart page.

NOTE: It doesn't work but am committing just to save progress. Also using a different branch so that I can still show the cart page design in the other branch.